backend/requirements.txt
# Requirements for the backend app

# Core web framework
fastapi==0.115.0
uvicorn==0.31.1

# Database
SQLAlchemy==2.0.23

# Schemas
pydantic==2.7.4
pydantic-settings==2.2.1

# Authentication / JWT
PyJWT==2.9.0
passlib[bcrypt]==1.7.4

# File uploads
python-multipart==0.0.9

# Environment / LLM / HTTP
python-dotenv==1.0.0
litellm
requests==2.32.3
reportlab==4.2.5
